#AB617 added gui package, refactored gui classes to gui package
