--- src/main/java/com/starboard/Main.java
@@ -1,36 +1,0 @@
-package com.starboard;//package sample;
-
-//import com.starboard.App;
-import com.starboard.App;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-import java.io.IOException;
-
-public class Main extends Application {
-
-    private static Scene scene;
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage primaryStage) throws Exception{
-        scene = new Scene(loadFXML("introScene"), 1000, 500);
-        primaryStage.setTitle("Star Board - First Blood");
-        primaryStage.setScene(scene);
-        primaryStage.show();
-    }
-    static void setRoot(String fxml) throws IOException {
-        scene.setRoot(loadFXML(fxml));
-    }
-
-    static Parent loadFXML(String fxml) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource(fxml + ".fxml"));
-        return fxmlLoader.load();
-    }
-}
