+++ src/main/java/module-info.java
@@ -1,12 +1,10 @@
-module com.starboard{
+module com.gui{
     requires javafx.controls;
     requires javafx.fxml;
     requires com.fasterxml.jackson.databind;
     requires java.desktop;
 
-    opens com.starboard to javafx.fxml, com.fasterxml.jackson.databind, javafx.graphics;
-    exports com.starboard;
-
-    exports com.starboard.items to com.fasterxml.jackson.databind;
+    opens com.gui to javafx.fxml, com.fasterxml.jackson.databind, javafx.graphics;
+    exports com.gui;
     opens com.starboard.items to com.fasterxml.jackson.databind;
 }
