+++ src/main/resources/com/gui/introScene.fxml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<BorderPane id="BorderPane" fx:id="introScene" styleClass="bodybg" stylesheets="@style.css" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/11.0.1" fx:controller="com.gui.ControllerIntro">
+    <center>
+        <VBox alignment="CENTER" spacing="10" style="-fx-border-color: blue; -fx-border-width: 3; -fx-border-style: dashed" xmlns:fx="http://javafx.com/fxml">
+            <padding>
+                <Insets bottom="10" left="10" right="10" />
+            </padding>
+            <Button fx:id="btnNewGame" onAction="#loadNewGame" prefWidth="90" text="New Game" />
+            <Button fx:id="btnSavedGame" prefWidth="90" text="Saved Game" />
+            <Button fx:id="btnExitGame" prefWidth="90" text="Exit" />
+        </VBox>
+
+
+    </center>
+</BorderPane>
